Align add-on type dropdown with original options
